requirements.txt
asgiref==3.11.1
Django==5.2.11
django-admin-sortable2==2.3.1
django-jazzmin==3.0.3
django-ratelimit==4.1.0
django-redis==6.0.0
django-tinymce==5.0.0
gunicorn==25.0.1
packaging==26.0
pillow==12.1.1
psycopg2-binary==2.9.11
python-dotenv==1.2.1
redis==7.2.1
sqlparse==0.5.5
whitenoise==6.11.0
